evaluator: change the remaining string functions.
Change the remaining string functions.
