+++ evaluator/builtin.go
@@ -53,15 +53,6 @@ var OldFuncs = map[string]OldFunc{
 	"ifnull": {builtinIfNull, 2, 2, true, false},
 	"nullif": {builtinNullIf, 2, 2, true, false},
 
-	// string functions
-	"replace":         {builtinReplace, 3, 3, true, false},
-	"strcmp":          {builtinStrcmp, 2, 2, true, false},
-	"convert":         {builtinConvert, 2, 2, true, false},
-	"substring":       {builtinSubstring, 2, 3, true, false},
-	"substring_index": {builtinSubstringIndex, 3, 3, true, false},
-	"locate":          {builtinLocate, 2, 3, true, false},
-	"trim":            {builtinTrim, 1, 3, true, false},
-
 	// information functions
 	"current_user":  {builtinCurrentUser, 0, 0, false, false},
 	"database":      {builtinDatabase, 0, 0, false, false},
@@ -110,13 +101,20 @@ var Funcs = map[string]Func{
 	"date_arith":        {builtinDateArith, 3, 3},
 
 	// string functions
-	"concat":    {builtinConcat, 1, -1},
-	"concat_ws": {builtinConcatWS, 2, -1},
-	"left":      {builtinLeft, 2, 2},
-	"length":    {builtinLength, 1, 1},
-	"lower":     {builtinLower, 1, 1},
-	"repeat":    {builtinRepeat, 2, 2},
-	"upper":     {builtinUpper, 1, 1},
+	"concat":          {builtinConcat, 1, -1},
+	"concat_ws":       {builtinConcatWS, 2, -1},
+	"left":            {builtinLeft, 2, 2},
+	"length":          {builtinLength, 1, 1},
+	"lower":           {builtinLower, 1, 1},
+	"repeat":          {builtinRepeat, 2, 2},
+	"upper":           {builtinUpper, 1, 1},
+	"replace":         {builtinReplace, 3, 3},
+	"strcmp":          {builtinStrcmp, 2, 2},
+	"convert":         {builtinConvert, 2, 2},
+	"substring":       {builtinSubstring, 2, 3},
+	"substring_index": {builtinSubstringIndex, 3, 3},
+	"locate":          {builtinLocate, 2, 3},
+	"trim":            {builtinTrim, 1, 3},
 }
 
 // See: http://dev.mysql.com/doc/refman/5.7/en/comparison-operators.html#function_coalesce
