+++ evaluator/builtin_string.go
@@ -180,103 +180,114 @@ func builtinUpper(args []types.Datum, _ context.Context) (d types.Datum, err err
 }
 
 // See: https://dev.mysql.com/doc/refman/5.7/en/string-comparison-functions.html
-func builtinStrcmp(args []interface{}, _ context.Context) (interface{}, error) {
-	if args[0] == nil || args[1] == nil {
-		return nil, nil
+func builtinStrcmp(args []types.Datum, _ context.Context) (d types.Datum, err error) {
+	if args[0].Kind() == types.KindNull || args[1].Kind() == types.KindNull {
+		d.SetNull()
+		return d, nil
 	}
-	left, err := types.ToString(args[0])
+	left, err := args[0].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
-	right, err := types.ToString(args[1])
+	right, err := args[1].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
 	res := types.CompareString(left, right)
-	return res, nil
+	d.SetInt64(int64(res))
+	return d, nil
 }
 
 // See: https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_replace
-func builtinReplace(args []interface{}, _ context.Context) (interface{}, error) {
+func builtinReplace(args []types.Datum, _ context.Context) (d types.Datum, err error) {
 	for _, arg := range args {
-		if arg == nil {
-			return nil, nil
+		if arg.Kind() == types.KindNull {
+			d.SetNull()
+			return d, nil
 		}
 	}
 
-	str, err := types.ToString(args[0])
+	str, err := args[0].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
-	oldStr, err := types.ToString(args[1])
+	oldStr, err := args[1].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
-	newStr, err := types.ToString(args[2])
+	newStr, err := args[2].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
+	d.SetString(strings.Replace(str, oldStr, newStr, -1))
 
-	return strings.Replace(str, oldStr, newStr, -1), nil
+	return d, nil
 }
 
 // See: https://dev.mysql.com/doc/refman/5.7/en/cast-functions.html#function_convert
-func builtinConvert(args []interface{}, _ context.Context) (interface{}, error) {
-	value := args[0]
-	Charset := args[1].(string)
-
+func builtinConvert(args []types.Datum, _ context.Context) (d types.Datum, err error) {
 	// Casting nil to any type returns nil
-	if value == nil {
-		return nil, nil
+	if args[0].Kind() != types.KindString {
+		d.SetNull()
+		return d, nil
 	}
-	str, ok := value.(string)
-	if !ok {
-		return nil, nil
-	}
+
+	str := args[0].GetString()
+	Charset := args[1].GetString()
+
 	if strings.ToLower(Charset) == "ascii" {
-		return value, nil
+		d.SetString(str)
+		return d, nil
 	} else if strings.ToLower(Charset) == "utf8mb4" {
-		return value, nil
+		d.SetString(str)
+		return d, nil
 	}
 
 	encoding, _ := charset.Lookup(Charset)
 	if encoding == nil {
-		return nil, errors.Errorf("unknown encoding: %s", Charset)
+		d.SetNull()
+		return d, errors.Errorf("unknown encoding: %s", Charset)
 	}
 
 	target, _, err := transform.String(encoding.NewDecoder(), str)
 	if err != nil {
+		d.SetNull()
 		log.Errorf("Convert %s to %s with error: %v", str, Charset, err)
-		return nil, errors.Trace(err)
+		return d, errors.Trace(err)
 	}
-	return target, nil
+	d.SetString(target)
+	return d, nil
 }
 
-func builtinSubstring(args []interface{}, _ context.Context) (interface{}, error) {
+func builtinSubstring(args []types.Datum, _ context.Context) (d types.Datum, err error) {
 	// The meaning of the elements of args.
 	// arg[0] -> StrExpr
 	// arg[1] -> Pos
 	// arg[2] -> Len (Optional)
-	str, err := types.ToString(args[0])
+	str, err := args[0].ToString()
 	if err != nil {
-		return nil, errors.Errorf("Substring invalid args, need string but get %T", args[0])
+		d.SetNull()
+		return d, errors.Errorf("Substring invalid args, need string but get %v", args[0].Kind())
 	}
 
-	t := args[1]
-	p, ok := t.(int64)
-	if !ok {
-		return nil, errors.Errorf("Substring invalid pos args, need int but get %T", t)
+	if args[1].Kind() != types.KindInt64 {
+		d.SetNull()
+		return d, errors.Errorf("Substring invalid pos args, need int but get %v", args[1].Kind())
 	}
-	pos := int(p)
+	pos := args[1].GetInt64()
 
-	length := -1
+	length := int64(-1)
 	if len(args) == 3 {
-		t = args[2]
-		p, ok = t.(int64)
-		if !ok {
-			return nil, errors.Errorf("Substring invalid pos args, need int but get %T", t)
+		if args[2].Kind() != types.KindInt64 {
+			d.SetNull()
+			return d, errors.Errorf("Substring invalid pos args, need int but get %v", args[2].Kind())
 		}
-		length = int(p)
+		length = args[2].GetInt64()
 	}
 	// The forms without a len argument return a substring from string str starting at position pos.
 	// The forms with a len argument return a substring len characters long from string str, starting at position pos.
@@ -284,48 +295,50 @@ func builtinSubstring(args []interface{}, _ context.Context) (interface{}, error
 	// In this case, the beginning of the substring is pos characters from the end of the string, rather than the beginning.
 	// A negative value may be used for pos in any of the forms of this function.
 	if pos < 0 {
-		pos = len(str) + pos
+		pos = int64(len(str)) + pos
 	} else {
 		pos--
 	}
-	if pos > len(str) || pos <= 0 {
-		pos = len(str)
+	if pos > int64(len(str)) || pos <= int64(0) {
+		pos = int64(len(str))
 	}
-	end := len(str)
-	if length != -1 {
+	end := int64(len(str))
+	if length != int64(-1) {
 		end = pos + length
 	}
-	if end > len(str) {
-		end = len(str)
+	if end > int64(len(str)) {
+		end = int64(len(str))
 	}
-	return str[pos:end], nil
+	d.SetString(str[pos:end])
+	return d, nil
 }
 
 // See: https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_substring-index
-func builtinSubstringIndex(args []interface{}, _ context.Context) (interface{}, error) {
+func builtinSubstringIndex(args []types.Datum, _ context.Context) (d types.Datum, err error) {
 	// The meaning of the elements of args.
 	// args[0] -> StrExpr
 	// args[1] -> Delim
 	// args[2] -> Count
-	fs := args[0]
-	str, err := types.ToString(fs)
+	str, err := args[0].ToString()
 	if err != nil {
-		return nil, errors.Errorf("Substring_Index invalid args, need string but get %T", fs)
+		d.SetNull()
+		return d, errors.Errorf("Substring_Index invalid args, need string but get %v", args[0].Kind())
 	}
 
-	t := args[1]
-	delim, err := types.ToString(t)
+	delim, err := args[1].ToString()
 	if err != nil {
-		return nil, errors.Errorf("Substring_Index invalid delim, need string but get %T", t)
+		d.SetNull()
+		return d, errors.Errorf("Substring_Index invalid delim, need string but get %v", args[1].Kind())
 	}
 	if len(delim) == 0 {
-		return "", nil
+		d.SetString("")
+		return d, nil
 	}
 
-	t = args[2]
-	c, err := types.ToInt64(t)
+	c, err := args[2].ToInt64()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
 	count := int(c)
 	strs := strings.Split(str, delim)
@@ -346,83 +359,92 @@ func builtinSubstringIndex(args []interface{}, _ context.Context) (interface{}, 
 		}
 	}
 	substrs := strs[start:end]
-	return strings.Join(substrs, delim), nil
+	d.SetString(strings.Join(substrs, delim))
+	return d, nil
 }
 
 // See: https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_locate
-func builtinLocate(args []interface{}, _ context.Context) (interface{}, error) {
+func builtinLocate(args []types.Datum, _ context.Context) (d types.Datum, err error) {
 	// The meaning of the elements of args.
 	// args[0] -> SubStr
 	// args[1] -> Str
 	// args[2] -> Pos
 	// eval str
-	fs := args[1]
-	if fs == nil {
-		return nil, nil
+	if args[1].Kind() == types.KindNull {
+		d.SetNull()
+		return d, nil
 	}
-	str, err := types.ToString(fs)
+	str, err := args[1].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
 	// eval substr
-	fs = args[0]
-	if fs == nil {
-		return nil, nil
+	if args[0].Kind() == types.KindNull {
+		d.SetNull()
+		return d, nil
 	}
-	subStr, err := types.ToString(fs)
+	subStr, err := args[0].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
 	// eval pos
 	pos := int64(0)
 	if len(args) == 3 {
-		t := args[2]
-		p, err := types.ToInt64(t)
+		p, err := args[2].ToInt64()
 		if err != nil {
-			return nil, errors.Trace(err)
+			d.SetNull()
+			return d, errors.Trace(err)
 		}
 		pos = p - 1
 		if pos < 0 || pos > int64(len(str)) {
-			return 0, nil
+			d.SetInt64(0)
+			return d, nil
 		}
 		if pos > int64(len(str)-len(subStr)) {
-			return 0, nil
+			d.SetInt64(0)
+			return d, nil
 		}
 	}
 	if len(subStr) == 0 {
-		return pos + 1, nil
+		d.SetInt64(pos + 1)
+		return d, nil
 	}
 	i := strings.Index(str[pos:], subStr)
 	if i == -1 {
-		return 0, nil
+		d.SetInt64(0)
+		return d, nil
 	}
-	return int64(i) + pos + 1, nil
+	d.SetInt64(int64(i) + pos + 1)
+	return d, nil
 }
 
 const spaceChars = "\n\t\r "
 
 // See: https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_trim
-func builtinTrim(args []interface{}, _ context.Context) (interface{}, error) {
+func builtinTrim(args []types.Datum, _ context.Context) (d types.Datum, err error) {
 	// args[0] -> Str
 	// args[1] -> RemStr
 	// args[2] -> Direction
 	// eval str
-	fs := args[0]
-	if fs == nil {
-		return nil, nil
+	if args[0].Kind() == types.KindNull {
+		d.SetNull()
+		return d, nil
 	}
-	str, err := types.ToString(fs)
+	str, err := args[0].ToString()
 	if err != nil {
-		return nil, errors.Trace(err)
+		d.SetNull()
+		return d, errors.Trace(err)
 	}
 	remstr := ""
 	// eval remstr
 	if len(args) > 1 {
-		fs = args[1]
-		if fs != nil {
-			remstr, err = types.ToString(fs)
+		if args[1].Kind() != types.KindNull {
+			remstr, err = args[1].ToString()
 			if err != nil {
-				return nil, errors.Trace(err)
+				d.SetNull()
+				return d, errors.Trace(err)
 			}
 		}
 	}
@@ -430,7 +452,7 @@ func builtinTrim(args []interface{}, _ context.Context) (interface{}, error) {
 	var result string
 	var direction ast.TrimDirectionType
 	if len(args) > 2 {
-		direction = args[2].(ast.TrimDirectionType)
+		direction = args[2].GetValue().(ast.TrimDirectionType)
 	} else {
 		direction = ast.TrimBothDefault
 	}
@@ -452,7 +474,8 @@ func builtinTrim(args []interface{}, _ context.Context) (interface{}, error) {
 	} else {
 		result = strings.Trim(str, spaceChars)
 	}
-	return result, nil
+	d.SetString(result)
+	return d, nil
 }
 
 func trimLeft(str, remstr string) string {
